pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sergey Radyukhin <22dollars@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
